+++ x/valset-pref/client/cli/query.go
@@ -3,7 +3,7 @@ package valsetprefcli
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/osmosis-labs/osmosis/v13/osmoutils/osmocli"
+	"github.com/osmosis-labs/osmosis/osmoutils/osmocli"
 	"github.com/osmosis-labs/osmosis/v13/x/valset-pref/client/queryproto"
 	"github.com/osmosis-labs/osmosis/v13/x/valset-pref/types"
 )