+++ x/valset-pref/types/msgs.go
@@ -5,7 +5,8 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/osmosis-labs/osmosis/v13/osmoutils"
+
+	"github.com/osmosis-labs/osmosis/osmoutils"
 )
 
 // constants